add kaitai files
"Kaitai struct" is a meta-language for defining the structure of binary files. the project has parsers that read .ksy files and generate file-parsers for that struct in any language you desire, such as python! this .py file is auto-generated. HOWEVER, the system does not have file WRITERS, so i don't expect to replace my custom reader/writer any time soon.